added euler string based pruning, seems very effective
