+++ src/distance/mod.rs
@@ -1,7 +1,11 @@
 mod graph;
 mod ids;
 mod nodes;
+mod str;
 mod tree;
+mod zs;
 
 pub use graph::{EClass, EGraph, ExtractionStats};
-pub use tree::{EditCosts, TreeNode, UnitCost, tree_distance, tree_distance_unit};
+pub use str::tree_distance_euler_bound;
+pub use tree::TreeNode;
+pub use zs::{EditCosts, UnitCost, tree_distance, tree_distance_unit};
